backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
instructor
openai
chromadb
python-dotenv